requirements.txt
boto3==1.17.112
python-dotenv==0.17.1
Scrapy==2.8.0
itemadapter==0.7.0